fix: Use Long instead of Int for maxRequestCharacters to prevent overflow - Convert DAO interface signatures from Int to Long - Update all implementations (Memory and File DAOs) - Update call site in ConversationImpl to pass Long directly - Prevents integer overflow when maxCacheCharacters exceeds Int.MAX_VALUE - Eliminates truncation of large character budgets for context window management

+++ aimo-core/src/main/kotlin/org/ivcode/aimo/core/conversation/ConversationImpl.kt
@@ -31,7 +31,7 @@ class ConversationImpl(
 
     override fun getMessages(maxCacheCharacters: Long?): List<AimoChatMessage>? {
         val entities = if (maxCacheCharacters != null && maxCacheCharacters > 0) {
-            conversationStore.getChatRequests(userId, chatId, maxCacheCharacters.toInt())
+            conversationStore.getChatRequests(userId, chatId, maxCacheCharacters)
         } else {
             conversationStore.getChatRequests(userId, chatId)
         }
